Update vim to not change cwd when opening a new file
I used to have it do this so that when I use fzf and want to start at
the directory that I'm current in, it is easier to do. However, I
realized if you select multiple results and then hit enter, those
results go into a Quickfix window. The problem is that you can't cycle
through those results because the Quickfix results show a full path to
the file, while the current directory is actually not at the root of the
project anymore.

Now instead of updating the cwd, just save the directory of the current
file to w:saved_dir. That way the current working directory is always
the root of the project, but I still have a variable that I can use when
I need to know the current file's directory (ex: <leader>A and <leader>F
mappings).