Add vim-tldr plugin and custom tldrrc file
The custom tldrrc file is mostly to specify my forked repo of
tldr-pages/tldr, because I have some custom pages in that repo that I
want to use (that aren't appropriate to PR to the main repo). This will
ensure that the node tldr app will get the pages that I want. Also in
the vim file, I make sure it points to my forked repo as well.