Set textwidth and colorcolumn on ft basis
Also set augroup for each filetype to remove/add the colorcolumn when
losing/gaining focus back on the filetype. Before, this was set
globally to col 81, so if a filetype used a different col when you leave
and then come back to the file, the colorcolumn would be set to the
wrong column.